+++ . npmignore
