+++ examples/composition-api/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "composition-api",
+  "name": "composition-api-example",
   "private": true,
   "version": "0.0.0",
   "type": "module",
@@ -10,7 +10,7 @@
   },
   "dependencies": {
     "vue": "^3.2.45",
-    "vue-window-size": "workspace:*"
+    "vue-window-size-composition-api": "workspace:*"
   },
   "devDependencies": {
     "@vitejs/plugin-vue": "^4.0.0",
