+++ package.json
@@ -24,6 +24,7 @@
     "happy-dom": "^8.1.4",
     "husky": "^8.0.3",
     "lint-staged": "^13.1.0",
+    "nx": "^15.5.1",
     "prettier": "^2.8.2",
     "typescript": "^4.9.4"
   },
