+++ packages/composition-api/package.json
@@ -1,0 +1,31 @@
+{
+  "name": "vue-window-size-composition-api",
+  "version": "1.0.0",
+  "main": "dist/index.js",
+  "module": "dist/index.mjs",
+  "types": "dist/index.d.ts",
+  "packageManager": "yarn@3.3.1",
+  "scripts": {
+    "build": "tsup",
+    "test:unit": "vitest --dir src --run",
+    "test:integration": "vitest --dir tests --run"
+  },
+  "devDependencies": {
+    "@types/node": "^18.11.18",
+    "@vitejs/plugin-vue": "^4.0.0",
+    "@vue/compiler-dom": "^3.2.45",
+    "@vue/runtime-core": "^3.2.45",
+    "@vue/test-utils": "^2.2.7",
+    "happy-dom": "^8.1.4",
+    "tsup": "^6.5.0",
+    "typescript": "^4.9.4",
+    "vite": "^4.0.4",
+    "vite-tsconfig-paths": "^4.0.3",
+    "vitest": "^0.27.1",
+    "vue": "^3.2.45"
+  },
+  "peerDependencies": {
+    "vue": ">=3.0.0",
+    "window-resize-subject": ">=1.5.0"
+  }
+}
