+++ packages/composition-api/src/composition-api.spec.ts
@@ -1,7 +1,7 @@
 import { describe, it, beforeEach, expect, vi } from 'vitest';
 import { defineComponent } from 'vue';
 import { shallowMount } from '@vue/test-utils';
-import { createUseWindowSize } from '~/composition-api';
+import { createUseWindowSize } from './composition-api';
 import type { WindowResizeSubject } from 'window-resize-subject';
 
 const mocks = {
