+++ packages/composition-api/src/index.ts
@@ -1,5 +1,5 @@
 import { createUseWindowSize } from './composition-api';
-import { getSubject } from './subject';
+import { getSubject } from '~shared/subject';
 
 // Composition API
 export const useWindowSize = createUseWindowSize(getSubject);
