+++ packages/mixin/tsconfig.json
@@ -1,0 +1,5 @@
+{
+  "extends": "../../tsconfig.json",
+  "include": ["src/**/*.ts", "src/**/*.tsx", "tests/**/*.ts"],
+  "references": [{ "path": "./tsconfig.node.json" }]
+}
