+++ packages/mixin/tsup.config.ts
@@ -1,0 +1,11 @@
+import { defineConfig } from 'tsup';
+
+export default defineConfig({
+  entry: ['src/index.ts'],
+  splitting: false,
+  sourcemap: false,
+  clean: true,
+  dts: true,
+  minify: true,
+  format: ['cjs', 'esm'],
+});
