+++ packages/plugin/tests/fixtures/OptionComponent.ts
@@ -1,0 +1,10 @@
+import { defineComponent } from 'vue';
+
+export default defineComponent({
+  template: `
+  <div>
+    <div id="width">{{ $windowWidth }}</div>
+    <div id="height">{{ $windowHeight }}</div>
+  </div>
+  `,
+});
