+++ packages/shared/src/config.ts
