+++ packages/shared/vite.config.ts
@@ -1,0 +1,12 @@
+/// <reference types="vitest" />
+
+import { defineConfig } from 'vite';
+import vue from '@vitejs/plugin-vue';
+import tsconfigPaths from 'vite-tsconfig-paths';
+
+export default defineConfig({
+  plugins: [vue(), tsconfigPaths()],
+  test: {
+    environment: 'happy-dom',
+  },
+});
