+++ tsconfig.json
@@ -11,6 +11,11 @@
     "esModuleInterop": true,
     "lib": ["ESNext", "DOM"],
     "skipLibCheck": true,
-    "noEmit": true
+    "noEmit": true,
+    "baseUrl": ".",
+    "paths": {
+      "~shared/*": ["./packages/shared/src/*"],
+      "~fixtures/*": ["./packages/shared/fixtures/*"]
+    }
   }
 }
